@@ -1,40 +1,41 @@
-# This file is used to configure your project.
-# Read more about the various options under:
-# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
-
 [metadata]
 name = pytileproj
 description = Class for working with TiledProjectionSystems (TPS)
 author = TU Wien
 author-email = remote.sensing@geo.tuwien.ac.at
-license = mit
-long-description = file: README.rst
-long-description-content-type = text/x-rst; charset=UTF-8
-url = https://github.com/pyscaffold/pyscaffold/
-project-urls =
+license = MIT
+license_files = LICENSE.txt
+long_description = file: README.rst
+long_description_content_type = text/x-rst; charset=UTF-8
+url = https://github.com/TUW-GEO/pytileproj
+project_urls =
     Documentation = https://pyscaffold.org/
-# Change if running only on Windows, Mac or Linux (comma-separated)
+
 platforms = any
+
 # Add here all kinds of additional classifiers as defined under
-# https://pypi.python.org/pypi?%3Aaction=list_classifiers
+# https://pypi.org/classifiers/
 classifiers =
     Development Status :: 4 - Beta
     Programming Language :: Python
 
 [options]
 zip_safe = False
-packages = find:
+packages = find_namespace:
 include_package_data = True
 package_dir =
     =src
-# DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
-setup_requires = pyscaffold>=3.3a0,<4
-# Add here dependencies of your project (semicolon/line-separated), e.g.
-# install_requires = numpy; scipy
-# The usage of test_requires is discouraged, see `Dependency Management` docs
-# tests_require = pytest; pytest-cov
-# Require a specific Python version, e.g. Python 2.7 or >= 3.4
-python_requires = >=3.6
+
+# Require a min/specific Python version (comma-separated conditions)
+# python_requires = >=3.8
+
+# Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
+# Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
+# new major versions. This works if the required packages follow Semantic Versioning.
+# For more information, check out https://semver.org/.
+install_requires =
+    importlib-metadata; python_version<"3.8"
+
 
 [options.packages.find]
 where = src
