@@ -1,4 +1,4 @@
-# Copyright (c) 2021, TU Wien, Department of Geodesy and Geoinformation
+# Copyright (c) 2025, TU Wien
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
